configured to laradock

+++ sample/components/vhost.py
@@ -136,7 +136,8 @@ def edit():
 
     if op != 'n':
         subprocess.getoutput(("rm " + env.CONFIG_DIR + edithost[1] + ".conf"))
-        subprocess.getoutput(("rm " + env.SITES_ENABLED + edithost[1] + ".conf"))
+        if SITES_ENABLED != false:
+            subprocess.getoutput(("rm " + env.SITES_ENABLED + edithost[1] + ".conf"))
         f = open(env.HOSTS_FILE, 'r', encoding="utf8")
         lines = f.readlines()
         f.close()
@@ -188,7 +189,8 @@ def delete():
 
     if op != 'n':
         subprocess.getoutput(("rm " + env.CONFIG_DIR + edithost[1] + ".conf"))
-        subprocess.getoutput(("rm " + env.SITES_ENABLED + edithost[1] + ".conf"))
+        if SITES_ENABLED != false:
+            subprocess.getoutput(("rm " + env.SITES_ENABLED + edithost[1] + ".conf"))
         f = open(env.HOSTS_FILE, 'r', encoding="utf8")
         lines = f.readlines()
         f.close()
